+++ CleverTapSDK/ABTesting/uieditor/messages/CTABTestEditorDeviceInfoMessageRequest.h
@@ -1,8 +1,0 @@
-#import "CTABTestEditorMessage.h"
-
-extern NSString *const CTABTestEditorDeviceInfoMessageRequestType;
-
-@interface CTABTestEditorDeviceInfoMessageRequest : CTABTestEditorMessage
-
-@end
-