+++ CleverTapSDK/CTInAppUtils.m
@@ -1,7 +1,8 @@
+
 #import <UIKit/UIKit.h>
 #import "CTInAppUtils.h"
 #if !CLEVERTAP_NO_INAPP_SUPPORT
-#import "CTInAppResources.h"
+#import "CTUIUtils.h"
 #endif
 
 static NSDictionary *_inAppTypeMap;
@@ -35,57 +36,32 @@ static NSDictionary *_inAppTypeMap;
 #if CLEVERTAP_NO_INAPP_SUPPORT
     return nil;
 #else
-    return [CTInAppResources bundle];
+    return [CTUIUtils bundle];
 #endif
 }
 
-+ (NSString *)XibNameForControllerName:(NSString *)controllerName {
-#if CLEVERTAP_NO_INAPP_SUPPORT
++ (NSString *)getXibNameForControllerName:(NSString *)controllerName {
+#if CLEVERTAP_NO_INAPP_SUPPORT || TARGET_OS_TV
     return nil;
-#else
-    return [CTInAppResources XibNameForControllerName:controllerName];
-#endif
-}
-
-+ (UIImage *)imageForName:(NSString *)name type:(NSString *)type {
-#if CLEVERTAP_NO_INAPP_SUPPORT
-    return nil;
-#else
-    return [CTInAppResources imageForName:name type:type];
-#endif
-    
-}
-
-+ (UIColor * _Nullable)ct_colorWithHexString:(NSString *)string {
-    
-    return  [self ct_colorWithHexString:string withAlpha:1.0];
-}
-
-+ (UIColor * _Nullable)ct_colorWithHexString:(NSString *)string withAlpha:(CGFloat)alpha {
-    
-    if (![string isKindOfClass:[NSString class]] || [string length] == 0) {
-        return [UIColor colorWithRed:0.0f green:0.0f blue:0.0f alpha:1.0f];
+#else    
+    NSMutableString *xib = [NSMutableString stringWithString:controllerName];
+    UIApplication *sharedApplication = [CTUIUtils getSharedApplication];
+    BOOL landscape = UIInterfaceOrientationIsLandscape(sharedApplication.statusBarOrientation);
+    if ([[UIDevice currentDevice] userInterfaceIdiom] == UIUserInterfaceIdiomPhone) {
+        if (landscape) {
+            [xib appendString:@"~iphoneland"];
+        } else {
+            [xib appendString:@"~iphoneport"];
+        }
+    } else {
+        if (landscape) {
+            [xib appendString:@"~ipadland"];
+        } else {
+            [xib appendString:@"~ipad"];
+        }
     }
-    
-    // Convert hex string to an integer
-    unsigned int hexint = 0;
-    
-    // Create scanner
-    NSScanner *scanner = [NSScanner scannerWithString:string];
-    
-    // Tell scanner to skip the # character
-    [scanner setCharactersToBeSkipped:[NSCharacterSet
-                                       characterSetWithCharactersInString:@"#"]];
-    [scanner scanHexInt:&hexint];
-    
-    // Create color object, specifying alpha
-    UIColor *color =
-    [UIColor colorWithRed:((CGFloat) ((hexint & 0xFF0000) >> 16))/255
-                    green:((CGFloat) ((hexint & 0xFF00) >> 8))/255
-                     blue:((CGFloat) (hexint & 0xFF))/255
-                    alpha:alpha];
-    
-    return color;
+    return [xib copy];
+#endif
 }
 
 @end