+++ CleverTapSDK/CleverTapBuildInfo.h
@@ -1,2 +1,2 @@
 
-#define WR_SDK_REVISION @"30902"
+#define WR_SDK_REVISION @"30903"