+++ CleverTapSDK/CleverTapJSInterface.h
@@ -1,5 +1,5 @@
-#if !(TARGET_OS_TV)
 #import <Foundation/Foundation.h>
+#if !(TARGET_OS_TV)
 #import <WebKit/WebKit.h>
 @class CleverTapInstanceConfig;