+++ CleverTapSDK/InApps/CTFooterViewController.m
@@ -14,7 +14,7 @@
 
 - (void)loadView {
     [super loadView];
-    [[CTInAppUtils bundle] loadNibNamed:[CTInAppUtils XibNameForControllerName:NSStringFromClass([CTFooterViewController class])] owner:self options:nil];
+    [[CTInAppUtils bundle] loadNibNamed:[CTInAppUtils getXibNameForControllerName:NSStringFromClass([CTFooterViewController class])] owner:self options:nil];
 }
 
 - (void)layoutNotification {