+++ CleverTapSDK/Inbox/models/CTInboxUtils.m
@@ -1,6 +1,7 @@
+
 #import "CTInboxUtils.h"
 #if !CLEVERTAP_NO_INBOX_SUPPORT
-#import "CTInAppResources.h"
+#import "CTUIUtils.h"
 #endif
 
 static NSDictionary *_inboxMessageTypeMap;
@@ -24,12 +25,12 @@ static NSDictionary *_inboxMessageTypeMap;
     return [_type integerValue];
 }
 
-+ (NSString *)XibNameForControllerName:(NSString *)controllerName {
++ (NSString *)getXibNameForControllerName:(NSString *)controllerName {
 #if CLEVERTAP_NO_INBOX_SUPPORT
     return nil;
 #else
     NSMutableString *xib = [NSMutableString stringWithString:controllerName];
-    UIApplication *sharedApplication = [CTInAppResources getSharedApplication];
+    UIApplication *sharedApplication = [CTUIUtils getSharedApplication];
     BOOL landscape = UIInterfaceOrientationIsLandscape(sharedApplication.statusBarOrientation);
     if (landscape) {
         [xib appendString:@"~land"];
@@ -37,6 +38,14 @@ static NSDictionary *_inboxMessageTypeMap;
         [xib appendString:@"~port"];
     }
     return xib;
+#endif
+}
+
++ (NSBundle *)bundle:(Class)bundleClass  {
+#if CLEVERTAP_NO_INBOX_SUPPORT
+    return nil;
+#else
+    return [CTUIUtils bundle];
 #endif
 }