+++ CleverTapSDK/include/CleverTap.h
@@ -1,0 +1,1 @@
+../CleverTap.h