+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+### [Version 3.9.3](https://github.com/CleverTap/clevertap-ios-sdk/releases/tag/3.9.3) (April 12, 2021)
+
+##### Added
+- Adds support for installation via Swift Package Manager
+- Addresses - https://github.com/CleverTap/clevertap-ios-sdk/issues/70
+
+##### Changed
+- Refactored code related to `CTTelephonyNetworkInfo` to address - https://github.com/CleverTap/clevertap-ios-sdk/issues/103
+
+##### Breaking 
+- Removes code related to Dynamic Variable and Visual Editor
+
 ### [Version 3.9.2](https://github.com/CleverTap/clevertap-ios-sdk/releases/tag/3.9.2) (February 5, 2021)
 
 ##### Fixed